requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
motor==3.5.1
pymongo==4.5.0
python-dotenv==1.0.1
openai==1.43.0
google-genai==0.6.0
google-auth==2.35.0
# Testes
pytest==8.3.3
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
